requirements.txt
apache-beam==2.19.0
avro-python3==1.9.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==0.21.24
fasteners==0.15
future==0.18.2
google-api-core==1.14.3
google-apitools==0.5.28
google-auth==1.7.1
google-cloud==0.34.0
google-cloud-bigquery==1.17.1
google-cloud-bigtable==1.0.0
google-cloud-core==1.0.3
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.25.0
hdfs==2.5.8
httplib2==0.12.0
idna==2.8
mock==2.0.0
monotonic==1.5
numpy==1.17.4
oauth2client==3.0.0
pbr==5.4.3
pkg-resources==0.0.0
protobuf==3.10.0
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydot==1.4.1
pymongo==3.9.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==3.13
requests==2.22.0
rsa==4.0
six==1.13.0
spotipy==2.4.4
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
